assignment-a02-architecting-and-governing-autonomous-ai-agents/app/requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
pyjwt==2.8.0
redis==4.5.5
sqlalchemy==2.0.22
pydantic==1.10.12
python-multipart==0.0.6        # for file uploads in FastAPI
aiofiles==23.1.0               # for async file reading
